requirements.txt
great-expectations==0.18.12
psycopg2-binary==2.9.9
pandas==2.2.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
pyarrow==15.0.0
